# doc/api.yaml
openapi: 3.0.4
info:
  title: WASAtext
  description:  API Web Chat App, University project
  version: 1.1.1
servers:
  - url: https://api.web-chat-app.com/v1
    description: Web Chat Application API Description
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  schemas:  
...
    Message:
      type: object
      description: Representation of a Message in the system
      properties:
        id:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Message Id
        contentMess:
          type: string
          minLength: 1
          maxLength: 5000
          pattern: '.*'             # gestire lato server
          description: Message content. May include text, emojis, links, and attachments.
        timestamp:
          type: string
          format: date-time
          minLength: 20
          maxLength: 50
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[\+\-]\d{2}:\d{2})$'
          description: Message timestamp
        statusInfo:
          type: string
          enum:
            - delivered
            - read
          description: Message status
        conversationId:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Conversation Id
        senderUserId:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: User ID who sent the message
        senderUsername:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of user who sent
          readOnly: true    # lo gestisco lato server, altrimenti se l'utente cambia nome, devo aggiornare tutti i messaggi precedenti --> l'username me lo ricavo lato server (dal DB)
        forwarded:                        # Indica se il messaggio è stato inoltrato da qualche altra chat --> Boolean 
          type: boolean
          default: false
          description: Indicates if message was forwarded
        reactions:
          type: array
          description: List of reactions
          minItems: 0
          maxItems: 300
          items:
            $ref: "#/components/schemas/Reaction"
      required: 
        - id
        - contentMess
        - timestamp
        - statusInfo
        - conversationId
        - senderUserId
        - senderUsername      # Lo calcolo con il JOIN nel DB -- se cambia username altrimenti avrei record non aggiornati
        - forwarded
        - reactions
    NewMessage:
      type: object
      description: Input payload for sending a new message. Server handles timestamp and author.
      properties:
        contentMess:
          type: string
          minLength: 1
          maxLength: 5000
          pattern: '.*'
          description: Message content (text, emojis).
      required:
        - contentMess
    MessageUpdate:
      type: object
      description: Input payload for editing a message.
      properties:
        contentMess:
          type: string
          minLength: 1
          maxLength: 5000
          pattern: '.*'
          description: The new content of the message.
      required:
        - contentMess
    ForwardMessage:                   # Inoltra un messaggio
      type: object
      description: Representation of a forwarded message
      properties:
        # messToForwardId: E' nell'URL del path
        targetConversationId:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Conversation to forward the message to (if known).
        targetUserId:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: User ID to forward the message to. (The server will find the existing chat or create a new one).

      #required: 
      #  - targetConversationId    # Opzionale -- nel caso non ho già una conversazione con quella chat
    Reaction:
      type: object
      description: Representation of a reaction to a message
      properties:
        reactionId: 
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Reaction ID
        emoji:
          type: string
          minLength: 1
          maxLength: 20  # Permette emoji composte 
          pattern: '.*'             # gestire lato server
          description: Emoji reaction (single or composite emoji)
        timestamp:
          type: string
          format: date-time
          minLength: 20
          maxLength: 50
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[\+\-]\d{2}:\d{2})$'
          description: Reaction timestamp
        messToReactId:         
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: The Messagge where will react           
        senderUserId:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: User who reacted
        senderUsername:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9._~-]+$'
          description: User username who reacted
          readOnly: true    # lo gestisco lato server, altrimenti se l'utente cambia nome, devo aggiornare tutti i messaggi precedenti --> l'username me lo ricavo lato server (dal DB)
      required: 
        - reactionId
        - emoji
        - senderUserId
        - senderUsername
...
    Error:
      type: object
      description: Textual representation of the error
      properties:
        error:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '.*'             
          description: Error message
        code:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '.*'             
          description: Error code
        details:
          type: string
          minLength: 0
          maxLength: 500
          pattern: '.*'             
          description: Additional error details
      required: 
        - error
  
## ENDPOINTS
paths:
...
  /wasatext/{username}/conversations/{conversationId}/messages:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
          minLength: 3                 
          maxLength: 30   
          pattern: '^[a-zA-Z0-9._~-]+$'
        description: Username of the user loggin in this session
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          minLength: 36     
          maxLength: 36      
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
        description: Unique Conversation ID
    
    get:                                                # ritorna i messaggi di una specifica conversazione
      operationId: getConversation
      summary: Get all messages in a conversation
      description: Returns messages in reverse chronological order (newest first)
      tags:
        - Messages
      parameters:
        - name: limit
          in: query
          description: Max number of messages to return (default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: before
          in: query
          description: Cursor timestamp. Returns messages OLDER than this time. Used for scrolling up.
          schema:
            type: string
            format: date-time
            minLength: 20
            maxLength: 50
            pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[\+\-]\d{2}:\d{2})$'
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                description: List of messages
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                    minItems: 0 # in ogni conversazione c'è almeno 1 messaggio (se una conversazione è iniziata, vuol dire che c'è almeno un messaggio --> per come l'ho implementata)
                    maxItems: 100
                    description: Messages in reverse chronological order in this Conversation
              example:
                messages:
                  - id: "d8f1a7c3-6e29-430a-bb42-73a92fe83d6c"
                    contentMess: "Hey everyone 👋 ready for tonight’s match?"
                    timestamp: "2025-11-05T10:25:43Z"
                    statusInfo: "read"
                    conversationId: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                    senderUserId: "a3f1c8b2-47e9-4db5-93a8-921ddbb512fc"
                    senderUsername: "luca_dev"    
                    forwarded: false
                    reactions: []

                  - id: "c8b6d99b-7d52-4a4d-b88a-f82f62221f31"
                    contentMess: "Absolutely! Let’s go ⚽🔥"
                    timestamp: "2025-11-05T10:27:12Z"
                    statusInfo: "delivered"
                    conversationId: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                    senderUserId: "a3f1c8b2-47e9-4db5-93a8-921ddbb512fc"
                    senderUsername: "luca_dev"    
                    forwarded: false
                    reactions: []
                  
        '403':
          description: User is not part of this conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"



    post:                                                                     # invio di un nuovo messaggio
      operationId: sendMessage
      summary: Send a new message
      description: Sending a new message in this conversation
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: "#/components/schemas/NewMessage"
      responses:
        '201':    # Created
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                id: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                contentMess: "Hey guys, I just created the group event for tonight 🎉"
                timestamp: "2025-11-05T10:48:31Z"
                statusInfo: "read"
                conversationId: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                senderUserId: "a3f1c8b2-47e9-4db5-93a8-921ddbb512fc"
                senderUsername: "luca_dev"    
                forwarded: false
                reactions: []
        '403':
          description: User is not part of this conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/conversations/{conversationId}/read_status:            # Leggo i messaggi ancora non letti in questa conversatione
    put:
      operationId: markMessagesAsRead
      summary: Mark messages as read
      description: |-
        Updates the status of messages to 'read'. 
        It marks the message specified in 'lastReadMessageId' and ALL preceding 
        unread messages in this conversation as 'read'.
        This is used to generate "blue ticks" for the sender.
      tags:
        - Messages
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30   
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user logging in this session
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36     
            maxLength: 36       
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Conversation ID
      requestBody:
        required: true
        description: The payload containing the ID of the last message seen by the user.
        content:
          application/json:
            schema:
              type: object
              description: Schema to define the last read message
              properties:
                lastReadMessageId:
                  type: string
                  format: uuid
                  minLength: 36     
                  maxLength: 36       
                  pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
                  description: The ID of the most recent message the user has visualized on screen.
              required:
                - lastReadMessageId
      responses:
        '200':  
          description: Status updated successfully
          content:
            application/json:
              schema:
                description: status update to "read"
                type: object
                properties:
                  success:
                    type: boolean
                    description: Successfully
                    example: true
              example:
                success: true
        '403':
          description: User is not part of this conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Conversation or Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/conversations/{conversationId}/messages/{messageId}:
    parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30  
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36     
            maxLength: 36      
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Conversation ID
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36     
            maxLength: 36      
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Message ID
  
    patch:                                                                        # Modifico un Messaggio
      operationId: editMessage
      summary: Edit a message content
      description: Update the text content of an existing message. Only the sender can edit.
      tags:
        - Messages
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              $ref: "#/components/schemas/MessageUpdate"
      responses:
        '200':
          description: Message updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                id: "c1f09a5d-44e2-4f7b-9a33-82dfcc19b4fe"
                contentMess: "Hey guys, change of plans — meeting at 8:30 instead of 8 ⏰"
                timestamp: "2025-11-05T11:10:45Z"
                statusInfo: "delivered"
                conversationId: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                senderUserId: "a3f1c8b2-47e9-4db5-93a8-921ddbb512fc"
                senderUsername: "filippo_dev"     
                forwarded: false
                reactions: []

        '403':
          description: User is not the sender of this message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


    delete:                                                   # Utente elimina un messaggio da lui inviato in quella conversazione
      operationId: deleteMessage
      summary: Delete a message
      description: Only the sender can delete their own messages
      tags:
        - Messages
      responses:
        '204':  # No Content
          description: Message deleted successfully
        '403':
          description: User is not the sender of this message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /wasatext/{username}/conversations/{conversationId}/messages/{messageId}/forward:
    post:                                                             # Inoltrare un messaggio
      operationId: forwardMessage
      summary: Forward a message to another conversation
      description: |-
        Creates a NEW message in the target conversation (specified in body) with the same content 
        as the original message. 
        The new message will have 'forwarded: true', a new timestamp (current time), 
        and the sender will be the current user (not the original author).
        -'targetConversationId': The message is appended to that specific chat.
        - 'targetUserId': if user choose a new private conversation The backend searches for a new private chat with this user.
      tags:
        - Messages
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3                 
            maxLength: 30              
            pattern: '^[a-zA-Z0-9._~-]+$'
          description: Username of the user loggin in this session
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36     
            maxLength: 36      
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Conversation ID
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            minLength: 36     
            maxLength: 36      
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: Unique Message ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardMessage"           
        
      responses:
        '201':
          description: Message forwarded successfully (New message created)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                id: "2e1f9b33-5c7a-4af8-b83e-9f12aa49d223"
                contentMess: "📸 Check out this photo!"
                timestamp: "2025-11-21T18:30:00Z"           # Timestamp aggiornato al momento in cui viene fatto il forward
                statusInfo: "delivered"
                conversationId: "e34f0b12-52a9-489c-87ce-a91e0fb4476d"
                senderUserId: "d8f1a7c3-6e29-430a-bb42-73a92fe83d6c"
                senderUsername: "filippo_dev"
                forwarded: true                             # Flag fondamentale da andare a cambiare nel backend
                reactions: []

        '400': 
          description: Invalid input (Must provide either targetConversationId or targetUserId)
          content:
            application/json:
              schema:
                 $ref: "#/components/schemas/Error"

        '403':
          description: User is not part of source or target conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Message or conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"





  #######################
  # ENDPOINTS REACTION
  #######################
...